refactoring and added trained dicts
